feat: выделить fetch из report
