+++ internal/collector/report/docs.go
@@ -1,12 +1,4 @@
-// Пакет занимается сбором и отправкой метрик на сервер. А так же незаметно регулирует значение счетчика pollCount.
-// Значение большинства метрик забираеются из пакета runtime.ReadMemStats() и ещё дополнительно пополняются двумя параметрами,
-// один из которых счетчик PollCount - там хранится число раз, сколько мы опросили память. После отправки на сервер данных, это
-// значение сбрасывается
-//
-// Основые функции:
-//
-// Fetch() - собирает основые параметры использования памяти и сохраняет
-// в промеждуточное хранилише типа Stats, а так же увеличивает счетчик PollCount
+// Пакет занимается отправкой метрик на сервер.
 //
 // Send() - отправляет ранее собранные метрики на сервер, и обнуляет значение счетчика PollCount
 // stats собирает информацию о использовании памяти, и сохраняет в хранилище
@@ -14,8 +6,8 @@
 //
 // Типичное использование:
 //
-// s := report.Fetch()
-// err := report.Send(s.ToTransport(),"http://my.serv.er/update")
+// ms := fetch.MemStats()
+// err := report.Send(ms.ToTransport(),"http://my.serv.er/update")
 //
 //	if err != nil{
 //		...
